third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/css-module/credentials.sub-expected.txt

This is a testharness.js-based test.
FAIL CSS Modules should be loaded with or without the credentials based on the same-origin-ness and the crossOrigin attribute assert_equals: Descendant CSS modules should be loaded with the credentials when the crossOrigin attribute is not specified and the target is same-origin expected (boolean) true but got (undefined) undefined
Harness: the test ran to completion.
